requirements.txt
pandas
pandas-gbq
requests
google-cloud-bigquery
